refactor: common 폴더 내부의 일관화(component/common/)
common 내부 디렉토리 구조를 기본적으로 전부 디렉토리로 두도록 설정

#107
