+++ FE/src/app/(admin)/admin/create/page.tsx
@@ -1,5 +1,5 @@
 import CreateForm from "@/components/common/form/program/CreateForm";
-import Title from "@/components/common/Title";
+import Title from "@/components/common/Title/Title";
 
 const ProgramCreatePage = () => {
   return (
