+++ FE/src/app/(admin)/admin/edit/[programId]/page.tsx
@@ -1,7 +1,7 @@
 //TODO: 서버 컴포넌트로 변경
 "use client";
 
-import Title from "@/components/common/Title";
+import Title from "@/components/common/Title/Title";
 import AccessRightValidate from "@/components/common/validate/AccessRight";
 import EditForm from "@/components/programEdit/EditForm";
