+++ FE/src/app/(admin)/admin/edit/error.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import ErrorFallback from "@/components/common/ErrorFallback";
+import ErrorFallback from "@/components/common/error/ErrorFallback";
 
 const EditPageError = () => {
   const error = {
