+++ FE/src/app/(auth)/login/name-error/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import Button from "@/components/common/Button";
-import ErrorFallback from "@/components/common/ErrorFallback";
+import Button from "@/components/common/Button/Button";
+import ErrorFallback from "@/components/common/error/ErrorFallback";
 import ERROR_CODE from "@/constants/ERROR_CODE";
 import ERROR_MESSAGE from "@/constants/ERROR_MESSAGE";
 import ROUTES from "@/constants/ROUTES";
