+++ FE/src/app/(private)/team-building/create/page.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from "next/navigation";
 import { toast } from "react-toastify";
 import LoadingSpinner from "@/components/common/LoadingSpinner";
-import Title from "@/components/common/Title";
+import Title from "@/components/common/Title/Title";
 import TeamBuildingCreateForm from "@/components/teamBuildingCreate/TeamBuildingCreateForm";
 import ERROR_CODE from "@/constants/ERROR_CODE";
 import ERROR_MESSAGE from "@/constants/ERROR_MESSAGE";
