+++ FE/src/components/common/Dropup.tsx
@@ -1,3 +1,4 @@
+//TODO: 사용하지 않는 컴포넌트
 import { PropsWithChildren, useState } from "react";
 import useOutsideRef from "@/hooks/useOutsideRef";
