+++ FE/src/components/common/error/ErrorFallback.tsx
@@ -1,7 +1,7 @@
 import { useQueryErrorResetBoundary } from "@tanstack/react-query";
 import Image from "next/image";
-import Button from "./Button";
-import Title from "./Title";
+import Button from "../Button/Button";
+import Title from "../Title/Title";
 
 const ERROR_TITLE = "ERROR";
 const RETRY_BUTTON_TEXT = "Try again";
