+++ FE/src/components/common/form/FormBtn.tsx
@@ -1,4 +1,4 @@
-import Button from "../Button";
+import Button from "../Button/Button";
 
 interface FormBtnProps {
   submitText: string;
