+++ FE/src/components/common/memberTable/compounds/ListInEditTypeProps.tsx
@@ -2,11 +2,11 @@ import classNames from "classnames";
 import { useContext, useState } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 import { toast } from "react-toastify";
-import CheckBox from "../../CheckBox";
+import CheckBox from "../../CheckBox/CheckBox";
+import ErrorFallback from "../../error/ErrorFallback";
 import AttendStatusToggle from "../../toggle/AttendStatusToggle";
 import MemberTableLoader from "../MemberTable.loader";
 import { MemberContext } from "../MemberTableWrapper";
-import ErrorFallback from "@/components/common/ErrorFallback";
 import ACTIVE_STATUS from "@/constants/ACTIVE_STATUS";
 import MESSAGE from "@/constants/MESSAGE";
 import { useGetProgramMembersByActive } from "@/hooks/query/useMemberQuery";
