+++ FE/src/components/manage/CancleBtn.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useRouter } from "next/navigation";
-import Button from "../common/Button";
+import Button from "../common/Button/Button";
 
 const CancleBtn = () => {
   const router = useRouter();
